Remove job name from product files & update GEMPAK module. (#3415)
As referenced in #290, the names of AWIPS WMO product files populating
the `atmos/wmo` output subdirectory have been featuring a portion of the
originating job script, `exgfs_atmos_awips_20km_1p0deg.sh`, in violation
of NCO code standards. In this PR, the offending `$job_name` variable is
removed, and the ecFlow definition of the `$job` variable relies only on
`$PBS_JOBNAME`.

Coincident to the above - and as referenced in #3300 - the GEMPAK
software module sourced on WCOSS will need to be updated to version
7.15.1 to mitigate the occurrence of a non-fatal error that has been
observed to add compute time needed for GEMPAK job completion as a
result of numerous error messages adding significant size to the
corresponding log file. The necessary one-line change is made to the
`versions/run.wcoss2.lua` file.

---------

Co-authored-by: christopher hill <christopher.m.hill@clogin04.cactus.wcoss2.ncep.noaa.gov>
Co-authored-by: christopher hill <christopher.m.hill@clogin02.cactus.wcoss2.ncep.noaa.gov>