Migration to role account for Jenkins on Hercules (#3423)
# Description

This PR has the necessary updates to the CI scripts in Jenkins for
running on **Hercules** with the glopara role account.

- Added some updates to the internal documentation to the ci platform
config files for clarification.
- Moved the directories: **HOMDIR**, **STMP**, **PTMP**, to `/work2
`since the role account on Hercules does not write to `/work`.

[PyGithub](https://github.com/PyGithub/PyGithub) is not in **Spack
Stack** yet and is needed in support of the GitHub logging and gist
creation python codes. This python module will be installed in
`$HOME/.local` in the role accounts until the module is added to **Spack
Stack**.

Same with **GitHub CLI** (which will be supported in Spack Stack 1.9) it
will be installed in `${ROLE_HOME}/utils/bin` which should be in role
account's `$PATH `where `ROLE_HOME` is at the same level as `fix` and
`data`.

This is setup to run in the role account with the following added
directories and utils:

`/work2/noaa/global/role-global/GFS_CI_CD/HERCULES/CI`

Location on disk where the CI test cases actually build and run
specified in
[Jenkinsfile](https://github.com/TerrenceMcGuinness-NOAA/global-workflow/blob/8cbb38c059a52342c982a9b8d7baf63c1d30dd1d/ci/Jenkinsfile#L9)

```
└── GFS_CI_CD
    ├── global-workflow
    │   └── workflow
    ├── GFS_BASH_CI
    ├── GitLab
    └── Jenkins
        ├── agent
        └── workspace

```
Top level CI directories (this will file structure be duplicated for
each role account)
```
GFS_CI_CD
```
CI system directory for Jenkins launching scripts and in support of BASH
CI
```
GFS_CI_CD/global-workflow
```
Directory where the JAVA agent is launched from
```
Jenkins/agent
```
Jenkins work space (system level workspace)
```
Jenkins/workspace
```


  Resolves #3378 

# Type of change
- [ ] Bug fix (fixes something broken)
- [ ] New feature (adds functionality)
- [x] Maintenance (code refactor, clean-up, new CI test, etc.)

# Change characteristics
<!-- Choose YES or NO from each of the following and delete the other
-->
- Is this a breaking change (a change in existing functionality)? NO
- Does this change require a documentation update? NO
- Does this change require an update to any of the following submodules?
NO (If YES, please add a link to any PRs that are pending.)
  - [ ] EMC verif-global <!-- NOAA-EMC/EMC_verif-global#1234 -->
  - [ ] GDAS <!-- NOAA-EMC/GDASApp#1234 -->
  - [ ] GFS-utils <!-- NOAA-EMC/gfs-utils#1234 -->
  - [ ] GSI <!-- NOAA-EMC/GSI#1234 -->
  - [ ] GSI-monitor <!-- NOAA-EMC/GSI-Monitor#1234 -->
  - [ ] GSI-utils <!-- NOAA-EMC/GSI-Utils#1234 -->
  - [ ] UFS-utils <!-- ufs-community/UFS_UTILS#1234 -->
  - [ ] UFS-weather-model <!-- ufs-community/ufs-weather-model#1234 -->
  - [ ] wxflow <!-- NOAA-EMC/wxflow#1234 -->

# How has this been tested?

This PR can be tested in place

---------

Co-authored-by: Terry McGUinness <terrene.mcguinness@gmail.com>
Co-authored-by: Terry McGuinness <terry.mcguinness@noaa.gov>
Co-authored-by: global Role-Account <role-global@hercules-login-1.hpc.msstate.edu>