Add unified jinja obs list to marine DA (#3530)
# Description

PR arranges to have marine analysis and LETKF tasks use the same jinja
obs list (list of observation sources) currently used by the prep ocean
obs task. This is to insure that the obs lists are consistent between
tasks and is a step towards moving towards the JCB configuration scheme.

Resolves https://github.com/NOAA-EMC/GDASApp/issues/1572


# Type of change
- [x] Maintenance (code refactor, clean-up, new CI test, etc.)

# Change characteristics

- Is this a breaking change (a change in existing functionality)? NO
- Does this change require a documentation update? NO
- Does this change require an update to any of the following submodules?
YES (If YES, please add a link to any PRs that are pending.)
  - [x] jcb-gdas: https://github.com/NOAA-EMC/jcb-gdas/pull/94
(a submodule of GDASApp)
# How has this been tested?

C48mx500_3DVarAOWCDA and C48mx500_hybAOWCDA GDASApp ctests on Hera and
Orion

# Checklist
- [x] Any dependent changes have been merged and published
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have documented my code, including function, input, and output
descriptions
- [x] My changes generate no new warnings
- [x] New and existing tests pass with my changes
- [x] This change is covered by an existing CI test or a new one has
been added
- [x] Any new scripts have been added to the .github/CODEOWNERS file
with owners
- [x] I have made corresponding changes to the system documentation if
necessary