Add support for 24hr-averaged ocn/ice products (#3535)
Add full capability to produce 24-hr ocean and ice products. In
particular, fixing a bug where the CICE history files are missing
when `cycle` is not `t00z` and when `FHOUT_ICE` is `24`. This fix has
been made possible since the changes from [UFSWM PR
#2437](https://github.com/ufs-community/ufs-weather-model/pull/2437)
have been merged and the GW PR #3145 updated the UFSWM hash to include
these UFSWM updates for global-workflow.

Resolves #2674