Update scripts to use the GW-RT global namelist (#3017)
This changes the ways the input.nml file is created. Instead of
creating a namelist, the variables that are set locally now use the
UFS `global_control.nml.IN` template and the UFS atparse function to
generate the namelist.

This should allow the GW branch to more easily test changes in the UFS
regression test framework. It will also allow GW to more easily
introduce UFS changes into GW.
 
The functionality in `parsing_namelists_FV3.sh` is mostly unchanged.
Since variables are still saved as `local`, the parsing of the template
will occur in this script. Sections in the namelist that were previously
not generated (stoch physics options for example) will be set and parsed
since the regression test template has these placeholders active. As
options are turned off and on, ! will be used in the namelist to comment
out sections that are not used.

The MOM6 variables have been moved here since the template place holders
are active in global_control.nml.IN.

Since the stoch physics options are active, the SKEB definitions were
also modified in the config.efcs files. If the first value in SKEB is
not -999., then the stoch physics subroutine will turn SKEB on (do_skeb
set to 1).

The FHZERO variable was also renamed to reflect what was in the namelist
template.