add dev folder to documentation for generate workflow script (#3705)
# Description
Minor update to documentation to note the new "dev" folder location for
ci and workflow directories in the continuous integration section. Note,
this is just in the section I noticed and was not a full search for
areas where there might need additional updates due to the new "dev"
folder.