bug fix for IAU: move restart frequency calculation (#3704)
This PR provides a bug fix for #3611 in which it has been found in
several cases for the restart output times to be out of sync between the
marine components and atmosphere. One reason is that the creation of the
array for restart output times occurred in forecast_predet.sh but it
used the "DO_IAU" variable as a trigger. However, "DO_IAU" value can
change in forecast_det if it is a restart run. Therefore, I have moved
the calculation to forecast_postdet and this showed tests like the
extended S2SW test properly started from the segment time instead of
restarting at the beginning.

Resolves https://github.com/NOAA-EMC/global-workflow/issues/3611