Updates for wave fix and wave downstream
This PR requires a wave fix update (see:
https://github.com/NOAA-EMC/global-workflow/issues/3669, completed),
which has the following updates:
* Updates to ww3_gint.WHTGRIDINT.bin files to have both an Input and an
output grid, this is possible since we no longer support multi-gird.
This along with wave changes mean that we can run wave downstream
product jobs even with the low resolution updates. Additional weights
files were added for the low res uglo_100km grid. This has several
related workflow changes here.
* Physics updates for waves for GFSv17 in ww3_grid.inp.uglo_15km to
improve wave performance
* Updates to multiple ww3_grid.inp to include "&OUTS USSP = 1, IUSSP =
3, STK_WN = 0.04, 0.110, 0.3305 /" which ensures wave -> ocean coupling
is active. This is particularly important for GEFSv13 grid, but this was
added to all ww3_grid.inp files to ensure that if output grids
eventually become an input grid this is not an issue.

This PR enables the dev/ci/cases/pr/C48_S2SW_extended.yaml test. 

There is a bug fix for restart times for IAU=NO, which is required to
get the second segment of the forecast to start half-way through for the
coupled model.
* Resolves https://github.com/NOAA-EMC/global-workflow/issues/3611 

Several points for both the low resolution and the high resolution grid
are not there and were causing the awips_bull job to fial for waves. The
bull_awips file has now been made resolution dependent. Several buoys
were removed to ensure the jobs succeeded. Following up to ensure those
are advertised on a PNS will be a seperate issue.

* Resolves https://github.com/NOAA-EMC/global-workflow/issues/3226

Updates were made to ensure that all wave gempak and downstream gridded
jobs were successful and match the output from v16. Updates were also
made to allow for both low and high resolution tests to output to the
same grid.
* Refs https://github.com/NOAA-EMC/global-workflow/issues/3156 
* Resolves https://github.com/NOAA-EMC/global-workflow/issues/3264 

This PR turns on wave boundary point calculations.