Update Fit2Obs to `v1.1.7.1` (#3800)
Update the version of the Fit2Obs package to the new `v1.1.7.1` tag.

This tag includes a bug fix for `COMIN_HISTORY` setting. To use this
new tag the workflow needs to add a new `COM_VRFYARCH` variable 
and update `fit2obs_ver=1.1.7.1`.

The new package has been installed on all supported platforms.

Refs #3799