Add check for prepbufr at end of prep job (#3816)
This PR adds a check for the existence of a prepbufr file to the end of
the prep job script. This check will help resolve a silent failure when
the prepbufr file generation fails but the error code does not get sent
up to the workflow level from the obsproc/prepobs package.

Resolves #691

---------

Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>