Update HPSS archiving so we can restart experiments from archive
This is the first step in ensuring we can restart retros from HPSS
archive. This PR makes sure we can restart an experiment from what we
have in archive. The next step will then be making sure we can replicate
an experiment (https://github.com/NOAA-EMC/GFS/issues/3 which this may
or may not already do) and lastly long term we are working on cleaning
up which files are archived the minimal amount possible. We will start
that by creating a list of files and what we think should be in HPSS
(https://github.com/NOAA-EMC/GFS/issues/4).

Fixes https://github.com/NOAA-EMC/global-workflow/issues/3764
Fixes https://github.com/NOAA-EMC/global-workflow/issues/3758
Also makes sure we save all of the restart files for MOM6 not just the
.res. file.