Fix bug in gempak walltime calculation (#3823)
This PR fixes a bug that inadvertently reset the resource settings for
the gempak metatask jobs. A line to obtain job resources was left during
work to add metatask functionality back in for post-processing jobs.
This lead to the walltime multiplication based on the # of forecast
hours to be undone. The gempak walltime is now also reduced from 30mins
to 5mins, which is the time needed for gempak to process a single
forecast hour. The metatask group multiplication now correctly increases
the walltime to 45mins (9 forecast hours x 5mins) and allows gempak the
needed time to complete successfully.

Changes:
- Remove line in `gfs_tasks.py` that reset resources (bug)
- Update gempak walltime to be 5mins per forecast hour instead of
30mins, with bug fix the walltime multiplication correctly works for the
gempak-per-forecast-hour jobs.
- Move `gpend` up in `exgfs_atmos_nawips.sh` to before the output file
is copied to COM. Also add err check for `gpend`, thanks @aerorahul !

Resolves #3630