Make aerosol DA optional for GCAFS (#3830)
This pull request adds in the option to 'cycle' for GCAFS but without
aerosol DA. Only the meteorology is updated through the offline analysis
job, and the aerosols are self-cycled. This functionality replicates
capabilities used in GEFSv12's aerosol member.

This also adds in a PR CI case without aerosol DA, and YAMLs to support
the GCAFSv1 configuration both with and without aerosol DA at C96 and
C384.

  Resolves #3785