Update Hera and GaeaC6 configuration for select JEDI jobs (#3868)
This modifies the configuration for the following jobs on the
following machines
- Hera:  aeroanlvar
- Gaea C6:  atmanlvarm, atmensanlobs, atmensanlsol

For each job on each machine the change reduces the number of tasks run
per node. it has been observed that some executables require more memory
after moving to spack-stack/1.9.2 This is the case for the above jobs on
the indicated machines.

Resolves #3862