Marine DA script hardening (#3797)
This hardens the JEDI marine analysis code by ensuring that if the
analysis observation statistics applications fails to initialize, run,
and produce the correct output, the `gdas_marineanlinit`,
`gdas_marineanlchkpt`, and `gdas_marineanlfinal` jobs won't fail and
crash the workflow.

While I was at it, I also cleaned up the error handling in `jedi.py`. I
eliminated some redundancies in how exceptions are captured and pass all
caught exceptions to the new "Workflow" exceptions created in
https://github.com/NOAA-EMC/wxflow/pull/61.

I've also updated the saving-to-com for JEDI marine and atmospheric DA
to save to COM_CONF. This required some updates to archiving.

Resolves NOAA-EMC/gdasapp#1738