Cleanup of symlinks for wxflow, etc., use of DATAoutput for model, update to triggers for wavepost jobs, and other misc. stuff (#3833)
This collects a few things.  Specifically it:
- removes softlinks to `wxflow` and relies on the setting of
`PYTHONPATH` if `wxflow` is not available as an importable module. This
will streamline when `wxflow` is part of a virtualenv or loaded via
spack-stack
- moves the links of `soca` and `ufsda` into `ush/python` which is added
to `PYTHONPATH`. This applies consistency for all importable python
modules
- uses `DATAoutput` for writing model output into. In particular, it
creates and writes to `FV3ATM_OUTPUT`, `MOM6_OUTPUT`, `CICE_OUTPUT` and
`WW3_OUTPUT`. In addition, WW3 restarts are written to `WW3_RESTART`
- uses `log.points` and `log.wave_grid` files for triggering wave
post-processing jobs instead of binary output from the next forecast
output time.

Resolves #3610 
Might also cover some of the other issues, but awaiting confirmation.