Update the gsi fix version for new reject list (#3891)
The reject list for GSI observations was updated to exclude many surface
stations for the coupled soil/atm DA in GFSv17. The new reject list was
added to the staged glopara fix directory under the timestamp `20250529`
with #3763. The global-workflow now needs to be updated to point to the
latest gsi fix version to make use of the new reject list.

Note: Issue #3795 originally intended to update the GSI hash in addition
to the gsi fix version, but the GSI hash was updated under a [separate
PR](https://github.com/NOAA-EMC/global-workflow/pull/3498) due to a
needed [bugfix](https://github.com/NOAA-EMC/GSI/pull/899).

Resolves #3795
Refs #3763