Preparing for using aerosol B rescaling from YAML with orography masks (#3939)
This PR lays the groundwork to support aerosol B matrix rescaling based
on YAML configuration rather than input fix files. This includes a
land/sea mask, which requires the orography files be included in the
runtime directory. This PR is opened first to simplify the process and
not require a coordinated merge with GDASApp.

  Resolves #3938