Update standalone global_cycle for coupled mode (#3850)
Currently when the model is run in coupled mode with MOM and CICE, the
standalone global_cycle utility still updates the foundation temperature
and sea ice in the restart files.  The global_cycle utility has been
updated to disable updating the foundation temperature and sea ice in the
restart files when the model is coupled. Note: it
will continue to update the lake points that are not included in the
ocean and sea ice models.

The majority of the needed changes are within the UFS_UTILS repo (from
@GeorgeGayno-NOAA). The workflow changes are only to pass the new
namelist parameter to the global_cycle utility and update the ufs_utils
hash.

Resolves #3522 
Refs https://github.com/ufs-community/UFS_UTILS/issues/1051