Reorganize parm/gdas (#4011)
This aims to finally organize the `parm/gdas` directory, which
currently has files distributed between Global Workflow and GDASApp,
since many of its directories are symlinked from the `parm` directory of
GDASApp, and move all files into GDASApp.

The files will be renamed in the GDASApp to give them a
unified naming convention. This then gives the environment variables
that point to these files a unified naming convention that corresponds
directly to that file name. This will make it easier for developers to
find the underlying file that these variables point to, a common issue
when working with JEDI tasks.