Update pathing to FV3 dir (now UFSATM) and update CDMBGWD var name (#4045)
As of
https://github.com/ufs-community/ufs-weather-model/commit/688de8eb922b37981450d91494cfb4de0bd780c5,
there are two changes that occurred in the UFS namelists and FV3 pathing
that need to be reflected in the Global Workflow scripts.

https://github.com/ufs-community/ufs-weather-model/pull/2788 : This PR
changed the namelist variable `CDMBWD` to `CDMBGWD`. This fix is just a
typo in the namelist that was finally corrected. No changes in model
results or behavior resulted from this fix.

https://github.com/ufs-community/ufs-weather-model/pull/2780
(https://github.com/ufs-community/ufs-weather-model/commit/d4bac232cd6126f285e717f40b12196849fbf747)
: When MPAS was added to the UFS, a change was made to the
directory/submodule naming. We decided to rename `./FV3/` to
`./UFSATM/`.
There are a few scripts where `./ufs_model.fd/FV3/` is referenced and
those will need to be changed to `./ufs_model.fd/UFSATM/`. Again, no
changes in model results or behavior should result from this update.

Closes #4014 
Closes #4017