Update j-jobs to be bash linter compliant (#4124)
Updates the j-job and pre-jjob bash scripts to be compliant with both
`shellcheck` and `shfmt`. This does *not* enforce some other standards
not checked by `shellcheck`.

The `.editorconfig` is also updated with guidelines for formatting bash
files (used by `shfmt`). Two different blocks with identical settings
had to be used to accommodate j-jobs, which have no file suffix.

Refs: #397