Update UFS Model  (#4138)
This updates the ufs-weather-model to the top of it's develop.
Additionally it adds in a few minor updates to resources (mostly this is
for machine variability/being full versus jobs consistently running
long).
Removes CHA which isn't a variable output in operations from WAV output.
Changes the MED pets based on scaling work by @dpsarmie and others.